requirements.txt
agentmemory
agentloop
easycompletion
agentbrowser
fuzzysearch
python-Levenshtein
prompt_toolkit
pyfiglet
PyPDF2
openai-whisper
yt_dlp
wtpsplit
fuzzywuzzy